# docs/swagger.yaml
definitions:
...
  model.AppSetting:
    properties:
      id:
        type: integer
      key:
        type: string
      value:
        type: string
    type: object
...
  model.ReqUpdateAppSettingByKey:
    properties:
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /appsetting/get-by-key/:key:
    get:
      operationId: GetByKey
      parameters:
      - default: '"nearby_radius"'
        description: ' '
        in: query
        name: key
        type: string
      responses:
        "200":
          description: 'data: model.AppSetting'
          schema:
            $ref: '#/definitions/model.AppSetting'
      security:
      - Token: []
      summary: Get App Setting Value (permission = all)
      tags:
      - AppSetting
  /appsetting/update-by-key/:key:
    post:
      operationId: UpdateByKey
      parameters:
      - default: '"nearby_radius"'
        description: ' '
        in: query
        name: key
        type: string
      - description: Update AppSetting
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.ReqUpdateAppSettingByKey'
      responses:
        "200":
          description: 'data: model.AppSetting'
          schema:
            $ref: '#/definitions/model.AppSetting'
      security:
      - Token: []
      summary: Update App Setting Value (permission = superadmin)
      tags:
      - AppSetting
...
swagger: "2.0"